Generate PDF file as release artifact
What else could be taken as a releasable artifact for markdown or html
source files than PDF files? Each iteration now is able to have an
"snapshot" of it by generating PDF "binaries".

Regarding the build tools, I picked Wkhtmltopdf over Pandoc, because
I found the latter quite limited for html to pdf conversion. By the way
madnight/docker-alpine-wkhtmltopdf Alpine based Docker image is a
great alternative in size ~105MB and it already comes patched to
not rely on a X11 server running inside the container.