Improve PDF content style and layout
The early iterations of this Resume had a "good enough" print
styles and layout having in mind that some web browsers won't let
to override the defaullt print CSS rules (Ex.: Firefox).

In this iteration I'm  adding a better set of print CSS rules and
some specific wkhtmltopdf ones. At the same time I've changed the
order in the content, based on it's apparent "relevance".