Update content (skills section) and layout
It took me a while to figure out that the skills sections is
mostly oriented to job recuiters, who most of the time just need
to check if a certain skill or knowledge exist. Based on that I
decided to make it shorter for the trees sake ;)

Two pages resume, yay!