Update style and layout
In older interations I was completely focus on generating a
"sparkling" html resume, overlooking the fact that printed versions
are more useful. So I deliverated to improve the PDF version by
doing small sacrifices in the html one, this could be balanced but
I'll need more time to work around it.