Use dagger for site and PDF generation
I needed try dagger on any of my projects. Granted, needing to dance
with CUE for the first time was not that enjoyable, since at a glance it
often don't report the actual "DSL" failure, instead it communicates
what happened in the implementation side which for a newcomer might be a
little too rough.

It is nice they are moving into a different approach with "cloak" and
I'm excited to see what comes around with it.

By the way, I took the opportunity to make build.sh script POSIX compliant,
due to portability concerns. Technically we shouldn't need "build
scripts" in dagger lands, but somethins plain old POSIX scripting felt
nicer and quicker #SorryNotSorry.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>