Groom resume styles
While I was updating my resume with 2022 information, I noticed that
several styles are off putting to say least. For example, the contact
table at the beginning was being forced to a fixed with, which is a nice
pandoc feature for scientific content but for this context is not that
useful.

Also, I updated the "release" script to not blow up when GH does not
allow it to push tags.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>