Apply noticeable refactor - less is more!
It is funny when you only meant to pull a misaligned thread on your
sweater and you end up reduing a whole part. Ok, bad analogy since my
knitting skills are quite bad, still you get my drif. The changes on
this commit can be listed as:

  * Use JS to reshuffle tables on mobiles
  * Removed the need for | hack
  * Disabled JS when printing pdf
  * Removed redundant CSS rules
  * Removed unused CSS rules
  * Removed unneeded markup
  * Joined Talks and Interests section

In a nutshell "less is more".

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>