Use column layout and update CI tooling
Every single time I looked at my CV I got uncomfortable knowing that
part of the "most relevant" data was beyond the fold, or even worse, on
the second page when printed.

Although I'm not fan of columns, I think this is the best use for them,
since recruiters and hiring managers tend to qualify/discard most
resumes at simple glance (sadly).

As a public disclosure, I'm not a designer by any means so any feedback
on the visual part is appreciated :)

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>