2024 Updates
I got most of these changes on draft during the whole 2024 year,
and I believe it is best to start the 2025 with this update.

The changes on this commit can be summarized as:

  - Improve wording on working experience
  - Update the latest talks badges
  - Remove a hack in the dagger file

I'm aware that there's still space for improvement, but I'd rather
push them on a separate iteration.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>