Reduce header clutter
So, with the last rewording the first page started to require too
much vertical space, which lead me to notice space for improvements
mostly in the header and contact table section.

Oh yeah, I'm aware there's still opportunity to improve the exp.
section but I'll push that to the next iteration.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>