Rephrase the experience w/bullets
Quite frankly, I'm surprised how it took me to realize that my
resume had information but required a great deal of "imagination"
to fill the gaps between the description and all the technologies
listed on the sidebar.

After that realization I had to jump in and rephrase my experience
so the readers has to guess less about me.

I'm aware that there's *always* place for improvements but at least
I'm at ease with this iteration.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>