Upgrade CI configurations
It's great to see how an older version of dagger (v0.2.x) was able
to hold tight through 2 years without any upgrade. However it was
time to the jump to the latest version, and I'm happy to say that
it was even easier than dealing with the cue file.

I also took the opportunity to upgrade all GH Actions that were
trivial, with the exception of one which I'm going to push for
future iterations.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>