Elaborate on development skills
I learned the hard way that my development side is not explicitly
mentioned, giving the impresion that I'm only either on the ops
or management side.

I think there's another way to achieve this, but for the meantime
I simply added the programming language name to the experience
bullets when possible.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>